nice lil refactoring
